requirements.txt
asgiref==3.5.2
Django==4.1.1
django-js-asset==2.0.0
django-mptt==0.14.0
sqlparse==0.4.3
django-debug-toolbar==3.7.0
